pyproject.toml
[project]
name = "cv-optimizer"
version = "0.1.0"
description = "Optimize CV for a given role using LLM (LiteLLM + LangChain)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-core>=0.3.0",
    "langchain-litellm>=0.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cv_optimizer", "src/utils"]